feat: updates

+++ TaskService/src/main/java/com/example/taskservice/Dto/TaskDto.java
@@ -7,7 +7,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -23,7 +22,7 @@ public class TaskDto {
     @NotNull
     private Long categoryId;
     @NotNull
-    private Double price;
+    private Float price;
     @NotBlank
     private String provider;
